backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-multipart==0.0.20
oracledb>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==4.0.1
